Amendments-2 | rename components
