+++ src/components/Lists.js
@@ -1,91 +1,24 @@
 import React, { Component } from 'react';
 
-import TextField from '@material-ui/core/TextField';
-import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-
 import { connect } from 'react-redux';
+import Item from './Item';
 import { addListAsync, getListsAsync } from '../actions/lists';
 
-const styles = theme => ({
-  root: {
-    width: '100%',
-    marginTop: theme.spacing.unit * 2,
-    overflowX: 'auto',
-  },
-});
 
 class Lists extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      title: '',
-      symbol: '',
-    };
-    this.handleButton = this.handleButton.bind(this);
-    this.handleChangeTitle = this.handleChangeTitle.bind(this);
-    this.handleChangeSymbol = this.handleChangeSymbol.bind(this);
-  }
-
   componentWillMount() {
     this.props.getLists();
   }
 
-  handleButton = () => {
-    this.props.addList({
-      title: this.state.title,
-      symbol: this.state.symbol,
-      id: new Date(),
-      items: [],
-    });
-    this.setState({ title: '', symbol: '', id: '' });
-  };
-
-  handleChangeTitle = (e) => {
-    this.setState({ title: e.target.value });
-  };
-
-  handleChangeSymbol = (e) => {
-    this.setState({ symbol: e.target.value });
-  };
-
   render() {
+    const { lists } = this.props;
     return (
-      <Paper className={styles.root}>
-        <Toolbar>
-          <div>
-            <Typography variant="title" id="tableTitle">
-              Add new List
-            </Typography>
-          </div>
-        </Toolbar>
-        <Grid container spacing={0} justify="center">
-          <Grid item xs={12} sm={3} />
-          <Grid item xs={12} sm={2}>
-            <TextField
-              label="Title"
-              value={this.state.title}
-              onChange={this.handleChangeTitle}
-              margin="normal"
-            />
-          </Grid>
-          <Grid item xs={12} sm={4}>
-            <TextField
-              label="Symbol"
-              value={this.state.symbol}
-              onChange={this.handleChangeSymbol}
-              margin="normal"
-            />
-            <Button onClick={this.handleButton} variant="raised" primary="true">
-              Submit
-            </Button>
-          </Grid>
-          <Grid item xs={12} sm={3} />
+      lists.map(list => (
+        <Grid item xs={12} lg={12} md={12} sm={12} key={list.id}>
+          <Item list={list} />
         </Grid>
-      </Paper>
+      ))
     );
   }
 }
